Android 9 support

+++ vport_backend/src/main/java/com/vport/model/Coordinate.java
@@ -4,29 +4,29 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class Coordinate {
-    private int x;
-    private int y;
+    private Float x;
+    private Float y;
 
     public Coordinate() {
     }
 
-    public int getX() {
+    public Float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(Float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public Float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(Float y) {
         this.y = y;
     }
 
-    public Coordinate(int x, int y) {
+    public Coordinate(Float x, Float y) {
         this.x = x;
         this.y = y;
     }
